Update posts.service.ts - allow html tags
Added <cite>, which I had forgotten. Also modified the allowable html code because there was a bug where the style attribute was not working as intended. Added allowStyles to hopefully fix this. It should allow any of the style options in that list for any tag for which the style attribute is permitted. <style> remains disallowed.