Move Bsky to settings, polish up some styles, fix some stuff
The renamed setting that I intended to put into the last commit is here now. additionally the tags in the muted words
section do not squish anymore, and bluesky is moved into it's own category in the settings. The bluesky page isn't *gone*
but it isn't any longer in the sidebar. There's probably some smaller changes I am forgetting in this summary