Fix frontend builds on docker
Bug was introduced in https://github.com/gabboman/wafrn/pull/376 as volumes are not available during build times, hence this was always building an empty image.

Fix is a bit cumbersome as volume mounts are only available during run time, so this would copy over the frontend parts to the volume at each startup.

Likely need a better way to get the "index.html" over to the backend service for SEO purposes though