Add support for daily backups
Adds support for a small management script that can backup&restore an instance. This contains:
* Database dump
* Uploads directory
* PDS Volume

It also makes sure to run a backup every day, and keep the last 10 days worth of backups. Backups are stored on local volume, people should add additional setup to move these to their preferred backup provider. An easy way might be to mount the backup directory as an S3 bucket and use any of the S3 compatible storage providers out there