Make the docker compose script safer
Due to a typo in the database volume mounts the database files have been added to an anonymous volume instead of a named one. This means if anyone is restarting the app using `docker compose down && docker compose up` they would get a new anonymous instance and seemingly nuke their database

This fixes to make sure that the named volume is mounted in the proper place.

However this will also seemingly cause the bug for anyone updating to this instance, so we might want to do some warning or no idea for those couple peoples who already using the compose scripts