Separate the media/cache domains in the default docker compose file
This separates the media and cache domains from the main app domain in the main docker compose file and caddyfiles.

This is a breaking change, anyone updating to this version will need to have a different app/media/cache domain setup otherwise Caddy will complain.

Change also makes sure that the media content is served through caddy, so no involvement from the backend is needed (although the cache backend is still used to fetch the media file, even for wafrn based content)