Add support for exporting user to ActivityPub backup
Adds support to export a user to an ActivityPub style backup. Resulting file should be compatible with the import script, which should import the posts, likes and bookmarks. Note that the blog's data will be exported, but the importer will still disregard it.

Both export and import only saves the posts of the user, and thread information is only linked in the backup, so during import it will only work if the resulting data is still available. Future idea might be supporting "full" backups where you not only backup the user's post, but the entire ancestor list as well, like how Tumblr does.

Export is done through the command line, but will put the resulting file to the media server so the user can be emailed manually about their backup. Cleanup of old backup files should be done manually

Main change to core backend code is the extraction of the `user->JSONLD` code from the route into a separate helper so it can be called from multiple places.