Fix security issues related to signature validation
Currently on POST requests we were not doing signature checks so an attacker could theoretically spam the sharedInbox with fake content.

Issue was likely due to the signature validation being disabled because the content-type was overridden at a middleware, which broke any signature where the service also had the content-type as part of the signature. I'm looking at you Mastodon

There's a bit of a refactor also going on to make the code path hopefully easier to understand, with explicit separate routes for GET and POST requests.

Also fixing a minor issue around signature requery - while we did support requerying the signature of users at least every 24 hours, that codepath was only working for POST requests, so only for people sending new content. Anyone trying to call the GET endpoints would remain stuck with their public key unchanged. This is updated by making sure we allow GET requests to intiate a re-query at least every 24 hours.

Also the caching for the keys didn't work, as the cache key was never set. This is now sorted, so hopefully there'll be a slight performance gain as well