atproto: be slightly more generous with image size
The Bluesky app's CDN will display images up to 2000 pixels, so
there's no need to crunch down to 768 as the first try. sharp
is capable of producing some reasonably small images within that
2000 pixel limit.

This tries resizing to 2000 pixels on the long end for atproto as
the first try, and only falls back to 768 pixels if that first
try still produced an image that's above 1000000 bytes.