Add support for Indieweb/Microformats compatible HTMLs
Modifies the backend generated HTML pages for the blog and post endpoints to return a microformats compatible representation of the relevant post / blog. On the blog site it will also return the latest 10 posts as well. Posts only contain the actual content, and not the reblog trail, but will use the same sanitization rules as the frontend (the logic was copied over as I didn't figure out a nice way yet to share code).

Since now we have proper content on the pages, the robots.txt is modified to allow the gathering of data from the fediverse/post endpoints. I'll add an option later so this can be turned on/off by the user.

Example posts, as seen by a bot:

* https://archive.is/T6QWm
* https://archive.is/PLXCR