Add support for backup cache providers
Since the entire wafrn stack is usually running on a single box, it has a single IP. When requesting resources from external sources it can hit IP limits easily.

This change will allow the bluesky/ATproto media cache function to use external backups. If the main request fails for any reason it will try one of the backup URLs as a fallback, and see if it can get the data through it.

An example implementation of the bacup provider is now built and deploted to Vercel. To configure it either natively:

```javascript
  externalCacheBackups: ['https://did-decoder-lambda.vercel.app/api/cache?media='],
```

or through the `.env` file (Note: the docker-compose.yml files have changed so they need to be updated to support this new env var):

```sh
FRONTEND_CACHE_BACKUP_URLS="'https://did-decoder-lambda.vercel.app/api/cache?media='"
```

The source code of the backup DID decoder is available at https://github.com/sztupy/did-decoder-lambda