optimizeMedia: preserve GIF losslessness sometimes
For huge GIFs it makes sense to convert them to video so they're not
eating up too much bandwidth, but GIFs are often used for things like
pixel art or other things where the fine details actually matter.
Crunching those down doesn't save much space, but exponentially reduces
the quality, so it's nice to preserve that if possible.

This updates the lossless check so it also produces lossless WEBPs if
the input file is a GIF under 2MB. Since wafrn's happy serving still
images that are 2MB, seems like an arbitrary but fair cutoff.